Icons adding capability for PasswordField only
Icons can be added to the left or right of PasswordField using setLeft(Node node) or setRight(Node node) methods.

Files changed are :
JFXPasswordField and JFXTextFieldSkin  to put an icon, for e.g. it can be put using FontAwesomeIconView

Note :-
The appropriate changes must be made to JFXTextField, that I haven't done, so please do it if you want.